Allow disabling hashie log when initializing client
This will save on log space and noise by preventing these kinds of
warnings from being dumped into the Rails log.
```
You are setting a key that conflicts with a built-in method Hashie::Mash#method defined in Kernel. This can cause unexpected behavior when accessing the key as a property. You can still access the key via the #[] method.
```